Remove all files
I don't actually want to remove these files. I'm just doing this so that
I can create a later commit which restores them. After that, I can open
a pull request which for the file restoration, which will allow me to
leverage GitHub's pull request feature to review the entire repo.

It's a weird workaround, but it'll work.